Added Tier 3

+++ src/SatisfactoryTree/Models/Recipe.cs
@@ -5,8 +5,8 @@
         public Recipe(
             decimal processingTimeInSeconds,
             decimal throughPutPerMinute,
-            Dictionary<string, int> inputs,
-            Dictionary<string, int> outputs,
+            Dictionary<string, decimal> inputs,
+            Dictionary<string, decimal> outputs,
             ManufactoringBuildingType manufactoringBuilding,
             ManufactoringMethodType manufactoringMethod = ManufactoringMethodType.Manufactured,
             bool primaryMethodOfManufacture = true)
@@ -21,8 +21,8 @@
         }
         public decimal ProcessingTimeInSeconds { get; internal set; }
         public decimal ThroughPutPerMinute { get; internal set; }
-        public Dictionary<string, int> Inputs { get; set; }
-        public Dictionary<string, int> Outputs { get; set; }
+        public Dictionary<string, decimal> Inputs { get; set; }
+        public Dictionary<string, decimal> Outputs { get; set; }
         public ManufactoringBuildingType ManufactoringBuilding { get; set; }
         public ManufactoringMethodType ManufactoringMethod { get; set; }
         public bool PrimaryMethodOfManufacture { get; set; }
